Finalizing Install Scripts
*** CHECKPOINT ***

~rbw

+++ data/bash_scripts/drop_all_moduli_generator_from_db.sh
@@ -1,7 +1,15 @@
 #!/usr/bin/env bash
-mariadb --defaults-file=/Users/ron/DOT_CNF/mg_admin.localhost.cnf <<EOF
-DROP DATABASE IF EXISTS moduli_db;
-DROP USER IF EXISTS 'moduli_generator'@'%';
-DROP USER IF EXISTS 'moduli_generator'@'localhost';
-FLUSH PRIVILEGES;
-EOF
+
+cnf={ ${1}: ~/DOT_CNF/mg_admin.localhost.cnf }
+	echo "Using MariaDB Configuration, "${cnf}"
+
+if [[ -f "${cnf}" ]]
+	mariadb --defaults-file=${cnf}<<EOF
+	DROP USER IF EXISTS 'moduli_generator'@'%';
+	DROP USER IF EXISTS 'moduli_generator'@'localhost';
+	FLUSH PRIVILEGES;
+	DROP DATABASE IF EXISTS moduli_db;
+	EOF
+
+else
+	echo "Invalid MariaDB, "${cnf}"
